Changed User's id element type to Long

+++ src/main/java/cstv/Interfaces/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("myUserRepository")
-public interface UserRepository extends MongoRepository<User, Integer> {
+public interface UserRepository extends MongoRepository<User, Long> {
 
     User findByUsername(String username);
 }
